fix: run AI catalog sync as standalone script to avoid DB dependency
The ci-ai-catalog-sync workflow was failing because the NestJS command
bootstrap tried to connect to PostgreSQL, which is not available in CI.
The sync command only fetches from models.dev and writes a JSON file —
it has zero database dependencies.

This converts it to a standalone ts-node script, removes the old NestJS
command and its module registration, and drops the unnecessary
twenty-server build step, making the workflow both functional and faster.

Made-with: Cursor
